16bit buffer support Add (non-black background) color initializer for (16 bit). For 16 bit colors 0x0000 is black For 32 bit colors 0x0000000 is transparent

+++ src/loaders/lottie/tvgLottieBuilder.cpp
@@ -1271,8 +1271,8 @@ void LottieBuilder::updateStrokeEffect(LottieLayer* layer, LottieFxStroke* effec
         }
     //A specific mask
     } else {
-        auto idx = static_cast<uint32_t>(effect->mask(frameNo) - 1);
-        if (idx < 0 || idx >= layer->masks.count) return;
+        auto idx = static_cast<int32_t>(effect->mask(frameNo) - 1);
+        if (idx < 0 || static_cast<uint32_t>(idx) >= layer->masks.count) return;
         layer->masks[idx]->pathset(frameNo, SHAPE(shape)->rs.path, nullptr, tween, exps);
     }
